# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedSupply B2B Pharmaceutical Distribution API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendor/products:
    get:
      operationId: listVendorProducts
      tags: [vendor]
      summary: List vendor's own products
      responses:
        "200":
          description: Vendor products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [vendor]
      summary: Upload a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created (pending verification)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendor/products/{id}:
    patch:
      operationId: updateVendorProduct
      tags: [vendor]
      summary: Update own product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVendorProduct
      tags: [vendor]
      summary: Delete own product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Product:
      type: object
      required: [id, vendorId, name, category, description, pricePerUnit, quantityAvailable, status, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        coaUrl:
          type: ["string", "null"]
        nafdacNumber:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        pricePerUnit:
          type: number
        quantityAvailable:
          type: integer
        status:
          type: string
          enum: [pending, verified, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, category, description, pricePerUnit, quantityAvailable]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        coaUrl:
          type: ["string", "null"]
        nafdacNumber:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        pricePerUnit:
          type: number
        quantityAvailable:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        coaUrl:
          type: ["string", "null"]
        nafdacNumber:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        pricePerUnit:
          type: number
        quantityAvailable:
          type: integer
...