# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Api
 * MedSupply B2B Pharmaceutical Distribution API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  category: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  coaUrl?: string | null;
  /** @nullable */
  nafdacNumber?: string | null;
  /** @nullable */
  barcode?: string | null;
  pricePerUnit: number;
  quantityAvailable: number;
}